pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixeldot"
version = "0.1.0"
description = "Token-efficient pixel art toolkit for AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["pixel-art", "ai-agent", "sprite", "claude", "token-efficient"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "Pillow>=9.0",
]

[project.optional-dependencies]
perf = [
    "numpy>=1.24",
]
dev = [
    "pytest>=7.0",
    "numpy>=1.24",
]

[project.urls]
Homepage = "https://github.com/WonsangYeon/pixeldot"
Repository = "https://github.com/WonsangYeon/pixeldot"
Issues = "https://github.com/WonsangYeon/pixeldot/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
